requirements.txt
pip==25.0.1
fastapi==0.115.11
uvicorn==0.30.3
validators==0.33.0
requests==2.32.4
pyyaml==6.0.2
pytz==2025.1
